+++ test/interpolation/features/definition/basic.test.ts
@@ -1,8 +1,6 @@
-import * as assert from 'assert';
-import * as vscode from 'vscode';
-import { showFile } from '../../../editorHelper';
 import { position, sameLineLocation } from '../../../util';
 import { getDocUri } from '../../path';
+import { testDefinition } from '../../../definitionHelper';
 
 describe('Should find definition', () => {
   const docUri = getDocUri('definition/Basic.vue');
@@ -22,16 +20,3 @@ describe('Should find definition', () => {
     await testDefinition(docUri, position(4, 5), sameLineLocation(tagUri, 0, 0, 0));
   });
 });
-
-async function testDefinition(docUri: vscode.Uri, position: vscode.Position, expectedLocation: vscode.Location) {
-  await showFile(docUri);
-
-  const result = (await vscode.commands.executeCommand(
-    'vscode.executeDefinitionProvider',
-    docUri,
-    position
-  )) as vscode.Location[];
-
-  assert.ok(result[0].range.isEqual(expectedLocation.range));
-  assert.equal(result[0].uri.fsPath, expectedLocation.uri.fsPath);
-}
