+++ test/interpolation/fixture/completion/classComponent/Child.vue
