+++ test/interpolation/fixture/definition/classComponent/Child.vue
@@ -1,0 +1,11 @@
+<template>
+  <div></div>
+</template>
+
+<script>
+import Vue from 'vue'
+import Component from 'vue-class-component'
+
+@Component({})
+export default class BasicClass extends Vue {}
+</script>
