+++ test/interpolation/fixture/definition/classComponent/Parent.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <child></child>
+    <child />
+  </div>
+</template>
+
+<script>
+import Vue from 'vue'
+import Child from './Child.vue'
+import Component from 'vue-class-component'
+
+@Component({
+  components: {
+    Child
+  },
+})
+export default class ParentClass extends Vue {}
+</script>
