+++ test/lsp/features/definition/basic.test.ts
@@ -1,6 +1,4 @@
-import * as assert from 'assert';
-import * as vscode from 'vscode';
-import { showFile } from '../../../editorHelper';
+import { testDefinition } from '../../../definitionHelper';
 import { location, position, sameLineLocation } from '../../../util';
 import { getDocUri } from '../../path';
 
@@ -26,16 +24,3 @@ describe('Should find definition', () => {
     await testDefinition(docUri, position(17, 7), location(itemUri, 5, 0, 7, 1));
   });
 });
-
-async function testDefinition(docUri: vscode.Uri, position: vscode.Position, expectedLocation: vscode.Location) {
-  await showFile(docUri);
-
-  const result = (await vscode.commands.executeCommand(
-    'vscode.executeDefinitionProvider',
-    docUri,
-    position
-  )) as vscode.Location[];
-
-  assert.ok(result[0].range.isEqual(expectedLocation.range));
-  assert.equal(result[0].uri.fsPath, expectedLocation.uri.fsPath);
-}
